lift imports to a plugin file
